add percentage validation for agent_fee and treasury_fee, use u8 instead of u64, introduce InvalidPercentage error

+++ contracts/croncat-tasks/src/tests/task_tests.rs
@@ -23,6 +23,7 @@ use croncat_sdk_tasks::{
 use cw20::Cw20ExecuteMsg;
 use cw_multi_test::{BankSudo, Executor};
 use cw_storage_plus::KeyDeserialize;
+use std::u8;
 
 use super::{
     contracts,
@@ -1030,8 +1031,8 @@ fn remove_tasks_with_queries_success() {
             cw20: None,
             coin: [Some(coin(5, DENOM)), None],
             gas: 50_000,
-            agent_fee: u64::default(),
-            treasury_fee: u64::default(),
+            agent_fee: u8::default(),
+            treasury_fee: u8::default(),
             gas_price: GasPrice::default(),
         },
     };
@@ -1058,8 +1059,8 @@ fn remove_tasks_with_queries_success() {
             cw20: None,
             coin: [Some(coin(5, DENOM)), None],
             gas: 50_000,
-            agent_fee: u64::default(),
-            treasury_fee: u64::default(),
+            agent_fee: u8::default(),
+            treasury_fee: u8::default(),
             gas_price: GasPrice::default(),
         },
     };
